move `TrajOptProblem` to own file
